pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "smartman"
version = "0.1.0"
description = "A modern, interactive TUI for Linux man pages with AI explanations and syntax highlighting."
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
keywords = ["man", "linux", "tui", "rich", "textual", "cli", "ai", "explanation"]
authors = [
    { name = "Yash Ambaskar", email = "ambaskaryash@gmail.com" },
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Shells",
    "Topic :: Documentation",
    "Topic :: Utilities",
]

dependencies = [
    "typer[all]",
    "rich",
    "textual[dev]",
    "pyyaml",
    "requests",
]

[project.urls]
Homepage = "https://github.com/ambaskaryash/smartman-cli"
Documentation = "https://github.com/ambaskaryash/smartman-cli#readme"
Repository = "https://github.com/ambaskaryash/smartman-cli.git"
Issues = "https://github.com/ambaskaryash/smartman-cli/issues"

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "textual-dev"]

[project.scripts]
smartman = "smartman.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["smartman"]
